Organize pages and dialogs folders in Celbridge.UserInterface
