+++ Celbridge/CoreServices/Celbridge.UserInterface/ViewModels/Dialogs/AlertDialogViewModel.cs
