+++ Celbridge/CoreServices/Celbridge.UserInterface/ViewModels/Dialogs/ConfirmationDialogViewModel.cs
