+++ Celbridge/CoreServices/Celbridge.UserInterface/ViewModels/Dialogs/InputTextDialogViewModel.cs
