+++ Celbridge/CoreServices/Celbridge.UserInterface/ViewModels/Dialogs/NewProjectDialogViewModel.cs
