+++ Celbridge/CoreServices/Celbridge.UserInterface/ViewModels/Dialogs/ProgressDialogViewModel.cs
