+++ Celbridge/CoreServices/Celbridge.UserInterface/ViewModels/Pages/SettingsPageViewModel.cs
@@ -1,7 +1,7 @@
-﻿namespace Celbridge.UserInterface.ViewModels;
+﻿namespace Celbridge.UserInterface.ViewModels.Pages;
 
 public partial class SettingsPageViewModel : ObservableObject
 {
     public SettingsPageViewModel()
-    {}
+    { }
 }
