+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Dialogs/AlertDialog.cs
