+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Dialogs/ConfirmationDialog.cs
