+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Dialogs/InputTextDialog.cs
