+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Dialogs/NewProjectDialog.cs
