+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Dialogs/ProgressDialog.cs
