+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Pages/EmptyPage.cs
