+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Pages/HomePage.xaml
