+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Pages/HomePage.xaml.cs
@@ -1,4 +1,5 @@
 using Celbridge.UserInterface.Models;
+using Celbridge.UserInterface.ViewModels.Pages;
 
 namespace Celbridge.UserInterface.Views;
