+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Pages/MainPage.cs
@@ -1,3 +1,4 @@
+using Celbridge.UserInterface.ViewModels.Pages;
 using Microsoft.UI.Input;
 using Windows.System;
 using Windows.UI.Core;
