+++ Celbridge/CoreServices/Celbridge.UserInterface/Views/Pages/TitleBar.xaml.cs
